Fastbill export, sending correct Customer Id
In Fastbill there is the Customer Number and the Customer Id. The Customer
Number is what we call "Fastbill Id" or "User Client Id". The Customer Id more
like an internal database id that is needed in API calls.

+++ lib/fastbill/invoice.go
@@ -12,7 +12,7 @@ const (
 
 type Invoice struct {
 	Id                   int64  `json:"INVOICE_ID,string,omitempty"`
-	CustomerId           int    `json:"CUSTOMER_ID,string"`
+	CustomerId           int64  `json:"CUSTOMER_ID,string"`
 	CustomerCostcenterId int64  `json:"CUSTOMER_COSTCENTER_ID,string,omitempty"`
 	CurrencyCode         string `json:"CURRENCY_CODE,omitempty"`
 	TemplateId           int64  `json:"TEMPLATE_ID,string,omitempty"`
